requirements.txt
numpy==1.21.5
opencv_python==4.5.5.62
pandas==1.5.2
Pillow==9.5.0
tensorboardX==2.6
torch==1.7.1
torchvision==0.8.2
yacs==0.1.6
GDAL==3.0.2
